Add Compojure and http-kit
Compojure is the de facto standard routing library in Clojure, and I am
already familiar with it. There may be better options out there for a
more complex setup (such as sharing routes with ClojureScript or defining
true REST APIs), but that is not the focus at the moment. I want to get
something working quickly.

http-kit is a Clojure library for both HTTP client and server. I am not too
familiar with it but do have some experience; I think it will work fine.
Plus, it has few dependencies.